Some proposed changes...
Trey:

here are a few proposed changes. my weekend went off the rails, so I wasn't able to pull myself away for as long as I might have liked to review this. BUT rather than getting nothing done, I figured I would at least do a little.
i am happy to provide some more comments in the next day OR so, but this is also probably a reasonable place for us to calibrate whether these comments provide benefit.

There a a few places with some substantive changes (but I tried to avoid that so that your voice rings true). having said that, why did I suggest these changes?

i felt that a couple of the paragraphs could be cleaned up OR tightened up in several ways:

* get to the point faster OR with less repetition
* present some ideas in a more positive light versus a negative light
* replaced a few phrases that may be less familiar to beginners with a more neutral term and the technical term in parentheses: something in the vein of "you may not get a change submitted (pull request merged)" VERSUS "you may not get a pull request merged" 

I hope this helps. Would be happy to suggest some additional changes if desired.

I think that near the middle/bottom of the doc, there are places where the verbiage could be tightened up to reduce the total word count and to increase the chance that readers will continue to the bottom.